TinyBlarney.Backends.Simulation.Types

Description

Documentation

type SimTime = Word64

A type representing a number of timesteps in a simulation

type Sample = (SimTime, Int)

A type representing the value of a Signal at a point in time

type Signal = [Sample]

A type representing a signal

type SignalMap = Map CircuitInterfacePath Signal

A dictionary of named Signals

type SimulationInputs

Arguments

= SignalMap
dictionary of input signals

The input signals to a simulation

type SimulationOutputs

Arguments

= SignalMap
dictionary of Output signals

The output signals from a simulation

data AllSimulationOutputs

The outputs from a simulation

Constructors

AllSimulationOutputs

Fields

effects :: [(SimTime, IO ())]

Stream of effects

signals :: SimulationOutputs

dictionary of output signals

type PureSimulator = SimulationInputs -> AllSimulationOutputs

A type defining the interface to a pure simulator

type Simulator = SimulationInputs -> IO SimulationOutputs

A type defining the interface to a simulator

class Simulatable a where

Methods

simulate :: a -> SimulationInputs -> IO SimulationOutputs