TinyBlarney.Backends.Simulation.Verilator

Description

Documentation

buildSimulatorWithVerilator :: Circuit -> IO Simulator

Build a Simulator for a given Circuit using verilator by first generating the verilog code and the appropriate C++ wrapper to then invoke verilator