TinyBlarney.Backends.Simulation.Verilator.Communication

Description

Documentation

data SimReq a

The type of requests sent to a TinyBlarney verilator simulation process

Constructors

SimReq

Fields

simCmd :: SimCommand

command to be performed

simTime :: SimTime

simulaton timestamp

payload :: a

payload for the command

type SimCommand = Word8

Type of commands sent to a TinyBlarney verilator simulation

pattern Evaluate :: SimCommand

Command to evaluate the verilator simulation with a given payload

pattern Finish :: SimCommand

Command to terminate the verilator simulation process

sendSimReq :: Storable a => Handle -> SimReq a -> IO ()

Send a single 'SimReq a' into the provided Handle to the input to a TinyBlarney verilator simulation process

sendSimReqs :: Storable a => Handle -> [SimReq a] -> IO ()

Send a sequence of 'SimReq a's into the provided Handle to the input to a TinyBlarney verilator simulation process

data SimRsp a

The type of responses received from a TinyBlarney verilator simulation

Constructors

SimRsp

Fields

simRes :: SimResult

returned result type

simTime :: SimTime

simulaton timestamp

payload :: a

payload for the result

type SimResult = Word8

Type of results received from a TinyBlarney verilator simulation

pattern Evaluated :: SimResult

Result indicating the evaluated output signals for the last simulation step

pattern Finished :: SimResult

Result indicating the simulation terminated

pattern Unknown :: SimResult

Result indicating an unknown command

receiveSimRsps :: Storable a => Handle -> IO [SimRsp a]

Receive a sequence of 'SimRsp a's from the provided Handle to the output from a TinyBlarney verilator simulation process