TinyBlarney.Core.BasicTypes

Description

This module provides basic types, helpers and other primitives used throughout TinyBlarney.

Documentation

module TinyBlarney.Core.CircuitInterface

types

data Primitive

Available primitive operations.

Constructors

Constant Integer BitWidth
Constant k w: a w-sized constant value k

inputs

no input

outputs

the w-bit constant value k

DontCare BitWidth
DontCare w: a w-sized don't care value

inputs

no input

outputs

the w-bit don't care value

And BitWidth
And w: a w-sized bitwise "and" of 2 operands

inputs

[x, y], 2 w-bit operands

outputs

the w-bit bitwise "and" of x and y

Or BitWidth
Or w: a w-sized bitwise "or" of 2 operands

inputs

[x, y], 2 w-bit operands

outputs

the w-bit bitwise "or" of x and y

Xor BitWidth
Xor w: a w-sized bitwise "xor" of 2 operands

inputs

[x, y], 2 w-bit operands

outputs

the w-bit bitwise "xor" of x and y

Invert BitWidth
Invert w: a ones' complement of an operand

inputs

x, a w-bit operand

outputs

the one's complement of x

Concatenate BitWidth BitWidth
Concatenate wx wy: a wx+wy-sized concatenation of 2 operands

inputs

[x, y], a wx-bit and a wy-bit x and y operands

outputs

the wx+wy-bit concatenation of x and y

Slice (Int, Int) BitWidth
Slice (hi, lo) w: a (hi-lo+1)-bit wide slice of a w-bit operand

inputs

x, a w-bit operand

outputs

the (hi-lo+1)-bit wide slice of x between bit indices hi and lo (both included)

Merge MergeStrategy Int BitWidth
Merge mStrat n w represents a merging primitive with the mStrat merging strategy, n pairs of 1-bit enables and associated w-wide inputs, and one w-wide output

inputs

[en0, in0, en1, in1, ...], n pairs of enN 1-bit enables and inN w-bit values

outputs

a single output, a w-bit value result of the merging of the n inputs according to mStrat

Register (Maybe Integer) BitWidth
Register mInit w represents a register with a potential initial value - if mInit == Nothing the initial value is dont care - if mInit == Just init the initial value is init

inputs

[x], a single w-bit value

outputs

a single output, the w-bit value init (or don't care) or the last written input value x

Custom Circuit
A custom component

Interface CircuitInterface
A circuit interface primitive. BVs with this primitive are flatten roots and/or flatten leaves (based on the direction of the port(s) described). The intended use for this primitive is to have exactly one instance of it per circuit, with a rich associated CircuitInterface. The CircuitInterface argument describes the primitive's interface as perceived by the rest of the circuit. The direction of the ports in this embedded CircuitInterface can be flipped to obtain the circuit's interface as perceived from the outside of the circuit (example: From a circuit environment's perspective (the context instanciating a circuit), an Interface with an input port A and an output port B describes a circuit which will consume values from its environment through its A port, and produce values through its B port. From within that circuit, the Interface is flipped, and presents A as an output port which produces values for the rest of the netlist to consume, and B as an input port which consumes values produced by the rest of the nets in the netlist. )

data MergeStrategy

Merging Strategy

Constructors

MStratOr

data Net

A type to represent a netlist node.

Constructors

Net

Fields

instanceId :: InstanceId

a unique instance identifier

primitive :: Primitive

a primitive operation

inputConnections :: [NetInputConnection]

a list of input connections

type NetPort = (InstanceId, CircuitInterfacePath)

A type to identify a Net port. NetPort is a '(InstanceId, CircuitInterfacePath)' pair.

type NetInput = NetPort

type NetOutput = NetPort

data NetConnection

A type to represent the connection to a Net input port. It holds information about the NetOutput on the other end of the connection if any and on inlined operations if any.

Constructors

NetConnection NetOutput
Constructor wrapping another Net's output.

NetConnectionInlined Primitive [NetConnection]
Constructor inlining a Primitive operation and a '[NetConnection]' list of its inputs.

type NetInputConnection = (CircuitInterfacePath, NetConnection)

A type to refer to an input port's connection. NetInputConnection is a '(CircuitInterfacePath, NetConnection)' pair.

type Netlist = Array InstanceId Net

A Netlist type synonym for 'Array InstanceId Net'.

data Circuit

A built Circuit with a name and a CircuitInterface, and possibly a Netlist or an implementation for a given backend

Constructors

Circuit

Fields

name :: String

circuit's name

interface :: CircuitInterface

circuit's interface

backingImplementation :: CircuitImplementation

backing implementation of a Circuit

data CircuitImplementation

Constructors

Netlist Netlist
a 'TinyBlarney Netlist

BackendFiles [(Backend, FilePath)]
a list of files for given backends

data Backend

A type to identify a TinyBlarney backend

Constructors

Verilog
TinyBlarney's verilog backend

pretty printers

prettyPrimitive :: Primitive -> Doc

Pretty print a Primitive.

prettyNet :: Net -> Doc

Pretty print a Net.

prettyNetPort :: NetPort -> Doc

Pretty print a NetPort.

prettyNetInput :: NetInput -> Doc

Pretty print a NetInput.

prettyNetOutput :: NetOutput -> Doc

Pretty print a NetOutput.

prettyNetConnection :: NetConnection -> Doc

Pretty print a NetConnection.

prettyNetInputConnection :: NetInputConnection -> Doc

Pretty print a NetInputConnection.

prettyCircuit :: Circuit -> Doc

Pretty-print a Circuit

prettyCircuitImplementation :: CircuitImplementation -> Doc

Pretty print a CircuitImplementation.

basic operations

getChildrenCircuits :: Circuit -> [Circuit]

Retreive children Circuits

getUniqueChildrenCircuits :: Circuit -> [Circuit]

Retreive children Circuits removing duplicates

getAllCircuits :: Circuit -> [Circuit]

Retreive all available nested Circuits in a Circuit

getAllUniqueCircuits :: Circuit -> [Circuit]

Retreive all available nested Circuits in a Circuit removing duplicates

primInterface :: Primitive -> CircuitInterface

Get the CircuitInterface of a Primitive.

primInputsInfo :: Primitive -> [(CircuitInterfacePath, BitWidth)]

Get general info about a Primitive's inputs.

primInputWidths :: Primitive -> [BitWidth]

Get the BitWidths of a Primitive's inputs.

primInputPaths :: Primitive -> [CircuitInterfacePath]

Get the CircuitInterfacePaths of a Primitive's inputs.

primOutputsInfo :: Primitive -> [(CircuitInterfacePath, BitWidth)]

Get general info about a Primitive's outputs.

primOutputWidths :: Primitive -> [BitWidth]

Get the BitWidths of a Primitive's outputs.

primOutputPaths :: Primitive -> [CircuitInterfacePath]

Get the CircuitInterfacePaths of a Primitive's outputs.

primPretty :: Primitive -> Doc

Pretty print a Primitive.

primEval :: Primitive -> [(CircuitInterfacePath, Integer)] -> [(CircuitInterfacePath, Integer)]

Evaluate a Primitive.

primEvalFirst :: Primitive -> [(CircuitInterfacePath, Integer)] -> Maybe Integer

Evaluate the first output of a Primitive.

ifcUnaryOp :: BitWidth -> BitWidth -> CircuitInterface

CircuitInterface for 1-input 1-output circuits a.k.a. unary op.

ifcBinaryOp :: BitWidth -> BitWidth -> BitWidth -> CircuitInterface

CircuitInterface for 2-inputs 1-output circuits a.k.a. binary op.