TinyBlarney.Core.Bit

Description

This module defines the 'Bit n' type and a set of primitive operations on that type.

Documentation

newtype Bit (n :: Nat)

Type representing a sized bit vector

Constructors

AsBit

Fields

bv :: BV

evaluateBit :: Bit n -> Maybe Integer

Evaluate a 'Bit n' to an Integer value if possible

unsafeEvaluateBit :: Bit n -> Integer

Evaluate a 'Bit n' to an Integer value or gives an error

bitNToInteger :: Bit n -> Maybe Integer

unsafeBitNToInteger :: Bit n -> Integer

bitNFromInteger :: KnownNat n => Integer -> Bit n

unsafeBitNFromInteger :: Integer -> Int -> Bit n

unsafeWidthOf :: Bit n -> Int

Get the size of a 'Bit n' without relying on its type

unsafeBitConstant :: Integer -> Int -> Bit n

Constant of an explicitly user-provided width

bitConstant :: forall n. KnownNat n => Integer -> Bit n

Constant of a given value

bitZero :: Bit 1

A single bit 0

unsafeBitZeros :: Int -> Bit n

All bits zero with explicitly user-provided width

bitZeros :: forall n. KnownNat n => Bit n

All bits zero

bitOne :: Bit 1

A single bit 1

unsafeBitOnes :: Int -> Bit n

All bits one with explicitly user-provided width

bitOnes :: forall n. KnownNat n => Bit n

All bits one

bitDontCare :: Bit 1

A single bit don't care

unsafeBitDontCares :: Int -> Bit n

All bits don't care with explicitly user-provided width

bitDontCares :: forall n. KnownNat n => Bit n

All bits don't care

bitAnd :: Bit n -> Bit n -> Bit n

bitAnd x y returns the bitwise "and" of x and y

bitOr :: Bit n -> Bit n -> Bit n

bitOr x y returns the bitwise "or" of x and y

bitXor :: Bit n -> Bit n -> Bit n

bitXor x y returns the bitwise "xor" of x and y

bitInvert :: Bit n -> Bit n

bitInvert x returns the ones' complement of x

bitConcat :: Bit n -> Bit m -> Bit (n + m)

bitConcat x y returns the concatenation of x above y

bitMerge :: MergeStrategy -> [(Bit 1, Bit n)] -> Bit n

bitMerge mStrat ins returns the signal resulting from merging all signals in ins according to their enable signal and the given MergeStrategy mStrat

bitDelay :: forall n. KnownNat n => Bit n -> Bit n

bitDelay x return the signal x delayed by one cycle

unsafeBitSlice :: forall n m. (Int, Int) -> Bit n -> Bit m

unsafeBitSlice (hi, lo) x returns the slice of (hi-lo+1)-bit wide slice of x between bit indices hi and lo (both included)

unsafeFromBitList :: [Bit 1] -> Bit n

Collapse a '[Bit 1]' bit list of size n to a 'Bit n' sized bit vector

unsafeToBitList :: Bit n -> [Bit 1]

Expand a sized bit vector 'Bit n' to a '[Bit 1]' bit list of size n