TinyBlarney.Core.Bits

Description

Documentation

class Bits a where

Minimal complete definition

Nothing

Associated Types

type SizeOf a :: Nat

Type level bit vector size of the type in Bits

type SizeOf a = SizeOf' (Rep a)

Methods

sizeOf :: a -> Int

Value level bit vector size of a value of the type in Bits

default sizeOf :: GBits a => a -> Int

pack :: a -> Bit (SizeOf a)

Convert to a bit vector

default pack :: GBits a => a -> Bit (SizeOf a)

unpack :: Bit (SizeOf a) -> a

Convert from a bit vector

default unpack :: GBits a => Bit (SizeOf a) -> a

getExternalInterface :: Maybe String -> a -> CircuitInterface

Get corresponding external port as a CircuitInterface

default getExternalInterface :: GBits a => Maybe String -> a -> CircuitInterface

toBVs :: a -> [BV]

Retrieve BVs from a type in Bits

default toBVs :: GBits a => a -> [BV]

fromBVs :: [BV] -> (a, Int, [BV])

Parse a value of type a in Bits from a '[BV]', and returns a triple with the value parsed first, the number of BVs consumed from the input list second, and the new remaining list third

default fromBVs :: GBits a => [BV] -> (a, Int, [BV])

externalInterface :: forall a. Bits a => Maybe String -> CircuitInterface

The "external" CircuitInterface for the type a in Bits

internalInterface :: forall a. Bits a => Maybe String -> CircuitInterface

The "internal" CircuitInterface for the type a in Bits