TinyBlarney.Core.BuildCircuit

Documentation

buildCircuitInterface :: BuildCircuitIfc t => t -> CircuitInterface

Build an interface for a circuit function based on its type signature

buildCircuitWith :: GenCircuit a => String -> CircuitInterface -> a -> Circuit

Build a Circuit from a name, a CircuitInterface and a TinyBlarney circuit function in GenCircuit

buildCircuit :: (BuildCircuitIfc a, GenCircuit a) => String -> a -> Circuit

Build a Circuit from a name and a TinyBlarney circuit function in BuildCircuitIfc and GenCircuit

customInstanceWithCircuit :: GenCircuit a => Circuit -> a

customInstanceWith :: (BuildCircuitIfc a, GenCircuit a) => String -> a -> a

customInstance :: (BuildCircuitIfc a, GenCircuit a) => String -> a