TinyBlarney.Core.CircuitInterface

Description

This module provides means to describe a circuit's interface. It is one of the most fundamental modules in TinyBlarney as the notion of a CircuiInterface is used to describe how TinyBlarney building blocks can interface together.

types

type InstanceId = Int

A type to represent a unique identifier. InstanceId is defined as Int.

type BitWidth = Int

A type to represent a width in bits. BitWidth is defined as Int.

data PortDir

A type to represent the direction of an interface port (In or Out).

Constructors

In

Out

data CircuitInterface

A type to represent the interface of a circuit.

Constructors

Product [CircuitInterface]
Compose a new CircuitInterface out of multiple CircuitInterfaces.

Meta MetaInfo CircuitInterface
Wrap a CircuitInterface with some MetaInfo.

Port PortDir BitWidth
Leaf CircuitInterface constructor, describing circuit port with its direction (PortDir) and width (BitWidth).

data MetaInfo where

A type to represent meta information on a CircuitInterface

Constructors

InstanceId :: InstanceId -> MetaInfo
Decorate a CircuitInterface with an InstanceId.

NameHint :: String -> MetaInfo
Decorate a CircuitInterface with a String serving as a hint when generating names (when generating verilog for example...).

DocString :: String -> MetaInfo
Decorate a CircuitInterface with a String serving as a general description or documentation of the interface.

Implicit :: String -> MetaInfo
Mark a CircuitInterface as being an "implicit" interface, meaning that the ports in this interface can be implicitly wired up. This is usefull for reset ports or clock ports for example.

data CircuitInterfacePath where

A type to describe a path through a CircuitInterface. It is useful to uniquely identify a specific sub-interface within a CircuitInterface, and specifically, a leaf sub-interface such as a Port. A CircuitInterfacePath is a sequence of integers ('Seq.Seq Int'), each representing a step through the tree of Product with the Int used as an index into the Product 's list of sub-interfaces. Meta nodes are traversed transparently (no extra step in the sequence).

Bundled Patterns

pattern (:<|) :: Int -> CircuitInterfacePath -> CircuitInterfacePath
A step followed by a path.

pattern (:|>) :: CircuitInterfacePath -> Int -> CircuitInterfacePath
A path followed by a step.

pattern NoStep :: CircuitInterfacePath
An empty path.

pattern Step :: Int -> CircuitInterfacePath
A single step.

type CircuitInterfaceQuery a = CircuitInterface -> Maybe a

A type representing a query on a CircuitInterface.

data CircuitLeafCtxt

An input context for a function to run on interface leaves.

Constructors

CircuitLeafCtxt

Fields

mInstanceId :: Maybe InstanceId

path :: CircuitInterfacePath

nameHints :: [String]

implicitTag :: Maybe String

ifc :: CircuitInterface

pretty printers

prettyCircuitInterface :: CircuitInterface -> Doc

Pretty print a CircuitInterface.

prettyCircuitInterfacePath :: CircuitInterfacePath -> Doc

Pretty print a CircuitInterfacePath.

basic circuit interface operations

metaInstanceId :: InstanceId -> CircuitInterface -> CircuitInterface

Wrap a CircuitInterface with an InstanceId.

metaNameHint :: String -> CircuitInterface -> CircuitInterface

Wrap a CircuitInterface with an name hint.

metaDocString :: String -> CircuitInterface -> CircuitInterface

Wrap a CircuitInterface with an name doc string.

metaImplicit :: String -> CircuitInterface -> CircuitInterface

Wrap a CircuitInterface as "implicit" (with a tag).

flipPortDir :: PortDir -> PortDir

Flip the direction in a PortDir.

flipCircuitInterface :: CircuitInterface -> CircuitInterface

Flip the direction of all ports in a CircuitInterface.

getPortOutBitWidth :: CircuitInterfaceQuery BitWidth

Get the width of a CircuitInterface if it is a port.

queryCircuitInterfaceAt :: CircuitInterfaceQuery a -> CircuitInterface -> CircuitInterfacePath -> Maybe a

Run a CircuitInterfaceQuery on a CircuitInterface at a given CircuitInterfacePath.

isCircuitInterfaceLeaf :: CircuitInterface -> Bool

Return True for a leaf CircuitInterface.

onCircuitInterfaceLeaves :: (CircuitLeafCtxt -> a) -> CircuitInterface -> [a]

Run a function on interface leaves receiving a CircuitLeafCtxt, and return a list of all results.

queryCircuitInterfaceLeaves :: CircuitInterfaceQuery a -> CircuitInterface -> [(CircuitInterfacePath, Maybe a)]

Run a CircuitInterfaceQuery on all the leaves of a CircuitInterface and return a list of all results together with their associated path in a list sorted on the paths.

circuit interface leaves queries

getImplicitPortIns :: CircuitInterface -> [(CircuitInterfacePath, CircuitInterface, String)]

Get all implicit input ports of a CircuitInterface.

getImplicitPortOuts :: CircuitInterface -> [(CircuitInterfacePath, CircuitInterface, String)]

Get all implicit output ports of a CircuitInterface.

getExplicitPorts :: CircuitInterface -> [(CircuitInterfacePath, CircuitInterface)]

Get all explicit ports of a CircuitInterface.

getExplicitPortInsInfo :: CircuitInterface -> [(CircuitInterfacePath, BitWidth)]

Get all explicit input ports information in a CircuitInterface.

getExplicitPortInPaths :: CircuitInterface -> [CircuitInterfacePath]

Get all explicit input ports paths in a CircuitInterface.

getExplicitPortInWidths :: CircuitInterface -> [BitWidth]

Get all explicit input ports widths in a CircuitInterface.

getExplicitPortOutsInfo :: CircuitInterface -> [(CircuitInterfacePath, BitWidth)]

Get all explicit output ports information in a CircuitInterface.

getExplicitPortOutPaths :: CircuitInterface -> [CircuitInterfacePath]

Get all explicit output ports paths in a CircuitInterface.

getExplicitPortOutWidths :: CircuitInterface -> [BitWidth]

Get all explicit output ports widths in a CircuitInterface.