TinyBlarney.Core.NetHelpers

Documentation

netInputPaths :: Net -> [CircuitInterfacePath]

get the input paths of a net

netInputs :: Net -> [NetInput]

get all NetInput input ports of a Net

netOutputsInfo :: Net -> [(CircuitInterfacePath, BitWidth)]

get the output info of a net

netOutputWidths :: Net -> [BitWidth]

get the output widths of a net

netOutputPaths :: Net -> [CircuitInterfacePath]

get the output paths of a net

netOutputs :: Net -> [NetOutput]

get all NetOutputs of a Net

netOutput :: Net -> NetOutput

get the first NetOutput of a Net

netPorts :: Net -> [NetPort]

get the NetPorts of a Net

netConnectionPorts :: NetConnection -> [NetPort]

get the NetPorts of a NetConnection

remapNetConnectionInstanceId :: (InstanceId -> InstanceId) -> NetConnection -> NetConnection

Helper function remap the InstanceIds of a NetConnection

remapNetInputConnectionInstanceId :: (InstanceId -> InstanceId) -> NetInputConnection -> NetInputConnection

Helper function remap the InstanceIds of a NetInput

remapNetInstanceId :: (InstanceId -> InstanceId) -> Net -> Net

Helper function remap the InstanceIds of a Net

netlistNamesWith :: ([String] -> Maybe String) -> Netlist -> Map NetPort String

Derive a map of names given a Netlist and a function to process name hints.

netlistNames :: Netlist -> Map NetPort String

Derive a map of names given a Netlist.

interfaceNamesWith :: ([String] -> Maybe String) -> Maybe String -> CircuitInterface -> Map CircuitInterfacePath String

Derive a map of names given a CircuitInterface and a function to process name hints.

interfaceNames :: Maybe String -> CircuitInterface -> Map CircuitInterfacePath String

Derive a map of names given a CircuitInterface.

defaultNameDerive :: [String] -> Maybe String

Default derivation of names from a collection of name hints

externalNetlistInterface :: Netlist -> (InstanceId, CircuitInterface)

Return the exposed external CircuitInterface of the given Circuit externalNetlistInterface :: Netlist -> CircuitInterface externalNetlistInterface nl = mconcat (snd $ ifcs) where ifcs = sortOn fst [ (nId, metaInstanceId nId $ flipCircuitInterface ifc) | n@Net{ instanceId = nId , primitive = Interface ifc } <- elems nl ]