TinyBlarney.Misc.Misc

Description

These functions do not depend on anything specific to TinyBlarney

Documentation

valueOf :: forall n. KnownNat n => Int

Lower a type of kind Nat to Int value (with type application @)

liftNat :: Integral a => a -> (forall n. KnownNat n => Proxy n -> k) -> k

Lift an integer value to type-level natural

clamp :: (Num a, Bits a) => Int -> a -> a

clamp a given value to the provided width

ceilDiv :: Integral a => a -> a -> a

integerToWord8List :: Integer -> [Word8]

word8ListToInteger :: [Word8] -> Integer

sizedListToInteger :: (Integral a, Integral b) => [(a, b)] -> Integer

sizeListIntegerToList :: (Integral a, Num b) => [a] -> Integer -> [b]

unfoldWhileM :: Monad m => (a -> Bool) -> m a -> m [a]

update :: [t] -> Int -> t -> [t]

transposePropagate :: forall a b. (Ord a, Bounded a) => [b] -> [[(a, b)]] -> [(a, [b])]